[Auth] Merge AuthenticateWithSessionAndCookie and LoginRequired
