+++ packages/cli/src/generate/specs/app/src/app/app.controller.ts
@@ -1,9 +1,7 @@
-import { AuthenticateWithSessionAndCookie, controller } from '@foal/core';
+import { controller } from '@foal/core';
 
 import { ViewController } from './controllers';
-import { User } from './entities';
 
-@AuthenticateWithSessionAndCookie(User)
 export class AppController {
   subControllers = [
     controller('/', ViewController),
