--- packages/core/src/auth/authenticate.hook.ts
@@ -1,20 +1,0 @@
-import { getManager } from 'typeorm';
-
-import { Class, Hook, HookDecorator } from '../core';
-import { AbstractUser } from './entities';
-
-export function AuthenticateWithSessionAndCookie(UserEntity: Class<AbstractUser>): HookDecorator {
-  return Hook(async ctx => {
-    if (!ctx.request.session) {
-      throw new Error('AuthenticateWithSessionAndCookie hook requires session management.');
-    }
-    if (!ctx.request.session.authentication || !ctx.request.session.authentication.hasOwnProperty('userId')) {
-      return;
-    }
-    ctx.user = await getManager().findOne(
-      UserEntity,
-      ctx.request.session.authentication.userId,
-      { relations: [ 'userPermissions', 'groups', 'groups.permissions' ] }
-    );
-  });
-}
