+++ packages/core/src/auth/index.ts
@@ -2,5 +2,4 @@ export * from './authentication';
 export * from './authorization';
 export * from './entities';
 export * from './utils';
-export * from './authenticate.hook';
 export { LoginRequired } from './login-required.hook';
