+++ packages/examples/src/app/app.controller.ts
@@ -1,12 +1,7 @@
-import {
-  AuthenticateWithSessionAndCookie,
-  controller,
-} from '@foal/core';
+import { controller } from '@foal/core';
 
 import { AuthController, ViewController } from './controllers';
-import { User } from './entities';
 
-@AuthenticateWithSessionAndCookie(User)
 export class AppController {
   subControllers = [
     controller('', AuthController),
