+++ packages/examples/src/app/controllers/view.controller.ts
@@ -1,9 +1,11 @@
 import { Context, Get, LoginRequired, PermissionRequired, render } from '@foal/core';
 
+import { User } from '../entities';
+
+@LoginRequired({ userEntity: User, redirect: '/login' })
 export class ViewController {
 
   @Get('/')
-  @LoginRequired({ redirect: '/login' })
   home(ctx: Context) {
     return render('./templates/home.html', {
       csrfToken: ctx.request.csrfToken()
